Move implementations to Internal namespaces
